requirements.txt
cairosvg>=2.8
numpy>=1.24
pillow>=10
scipy>=1.11
